Refactor and document asm helpers
