+++ semanticdb/integration/src/main/java/com/javacp/Javacp.java
@@ -14,12 +14,10 @@ public class Javacp<A extends CharSequence & Serializable, B> extends ArrayList<
     }
 
     public void wildcard(ArrayList<?> wildcard) {}
-
     public void wildcard(Collection<? extends Number> a, Collection<? super Number> b) {}
 
     class InnerShadowTypeParam<A> {
-        public void move(A a, B b) {
-        }
+        public void move(A a, B b) { }
     }
 
     ArrayList<A> genericField;
@@ -28,22 +26,16 @@ public class Javacp<A extends CharSequence & Serializable, B> extends ArrayList<
     ArrayList<A> genericMethod() {
         return null;
     }
-
     int[] arrayMethod() {
         return new int[0];
     }
+    void genericParams(A a, B b) { }
+    void primitiveParams(int a, long b, float c, double d, short e, byte f, boolean g, char h) { }
+    void typeParams(ArrayList<HashMap<A, String[]>> a, Hashtable<String, B> b) { }
+    <C extends Integer> void methodTypeParams(C c) { }
 
-    void genericParams(A a, B b) {
-    }
-
-    void primitiveParams(int a, long b, float c, double d, short e, byte f, boolean g, char h) {
-    }
-
-    public void overload(java.io.File a) {
-    }
-
-    public void overload(sourcecode.File a) {
-    }
+    public void overload(java.io.File a) { }
+    public void overload(sourcecode.File a) { }
 
     // primitive fields
     public int Int;
@@ -54,30 +46,17 @@ public class Javacp<A extends CharSequence & Serializable, B> extends ArrayList<
     public boolean Boolean;
     public char Char;
 
-    void typeParams(ArrayList<HashMap<A, String[]>> a, Hashtable<String, B> b) {
-    }
-
-    <C extends Integer> void methodTypeParams(C c) {
-    }
 
     private int privateField;
     protected int protectedField;
     public int publicField;
     int packagePrivateField;
 
-    private void privateMethod() {
-    }
+    private void privateMethod() { }
+    protected void protectedMethod() { }
+    public void publicMethod() { }
+    void packagePrivateMethod() { }
 
-    protected void protectedMethod() {
-    }
-
-    public void publicMethod() {
-    }
-
-    void packagePrivateMethod() {
-    }
-
-    public Serializable anon = new Serializable() {
-    };
+    public Serializable anonymous = new Serializable() { };
 
 }
