+++ semanticdb/integration/src/main/java/com/javacp/NonGeneric.java
@@ -5,7 +5,5 @@ import java.io.Serializable;
 
 public class NonGeneric extends java.lang.Exception implements Serializable, java.io.Flushable {
     @Override
-    public void flush() throws IOException {
-
-    }
+    public void flush() throws IOException { }
 }
