+++ semanticdb/integration/src/main/scala/example/Example.scala
@@ -1,9 +1,12 @@
 package example
 
+import scala.concurrent.Future
+
 object Example { self =>
   new scala.collection.mutable.Stack[Int]()
   def main(args: Array[String]): Unit = {
     println(1)
   }
   val x = scala.reflect.classTag[Int]
+
 }
