+++ semanticdb/metacp/src/main/scala/scala/meta/internal/metacp/Main.scala
@@ -5,6 +5,7 @@ import java.nio.file._
 import java.nio.file.attribute.BasicFileAttributes
 import java.util.Comparator
 import scala.collection.mutable
+import scala.meta.internal.metacp.asm.Scopes
 import scala.meta.internal.{semanticdb3 => s}
 import scala.meta.internal.semanticdb3.Accessibility.{Tag => a}
 import scala.meta.internal.semanticdb3.SymbolInformation.{Kind => k}
