+++ semanticdb/metacp/src/main/scala/scala/meta/internal/metacp/asm/FieldSignatureVisitor.scala
@@ -1,0 +1,3 @@
+package scala.meta.internal.metacp.asm
+
+class FieldSignatureVisitor extends JavaTypeSignatureVisitor(false)
