+++ semanticdb/metacp/src/main/scala/scala/meta/internal/metacp/asm/TypedSignatureVisitor.scala
@@ -1,0 +1,6 @@
+package scala.meta.internal.metacp.asm
+
+/** Helper to abstract over SignatureVisitors that produce a result of type T */
+abstract class TypedSignatureVisitor[+T] extends FailFastSignatureVisitor {
+  def result(): T
+}
