+++ tests/jvm/src/test/resources/metacp.expect
@@ -220,7 +220,7 @@ _root_.com.javacp.Javacp#[A] => typeparam A: <: CharSequence with Serializable
   Serializable => _root_.java.io.Serializable#
 _root_.com.javacp.Javacp#[B] => typeparam B: <: Object
   Object => _root_.java.lang.Object#
-_root_.com.javacp.Javacp#anon. => var anon: Serializable
+_root_.com.javacp.Javacp#anonymous. => var anonymous: Serializable
   Serializable => _root_.java.io.Serializable#
 _root_.com.javacp.Javacp#arrayField. => private[javacp] var arrayField: Array[Int]
   Array => _root_.scala.Array#
