+++ tests/jvm/src/test/resources/metacp.owners
@@ -48,7 +48,7 @@ _root_.com.javacp.Javacp#Long. => <none>
 _root_.com.javacp.Javacp#Short. => <none>
 _root_.com.javacp.Javacp#[A] => _root_.com.javacp.Javacp#
 _root_.com.javacp.Javacp#[B] => _root_.com.javacp.Javacp#
-_root_.com.javacp.Javacp#anon. => <none>
+_root_.com.javacp.Javacp#anonymous. => <none>
 _root_.com.javacp.Javacp#arrayField. => <none>
 _root_.com.javacp.Javacp#arrayMethod(). => <none>
 _root_.com.javacp.Javacp#compareTo(B). => <none>
